add e2e test for test testing
