+++ tests/e2e/fixtures/main_using_simple_function.cairo
@@ -1,0 +1,8 @@
+%lang starknet
+from my_lib.utils import get_my_number
+
+@view
+func my_func() -> (res : felt):
+    let (res) = get_my_number()
+    return (res)
+end
