+++ tests/e2e/fixtures/test_main_using_simple_function.cairo
@@ -1,0 +1,9 @@
+%lang starknet
+from src.main import my_func
+
+@view
+func test_my_func{syscall_ptr : felt*, range_check_ptr}():
+    let (res) = my_func()
+    assert res = 42
+    return ()
+end
